Moves Quiz to its own file in production
