+++ QuizEngineTests/QuizTest.swift
@@ -7,23 +7,7 @@
 
 import Foundation
 import XCTest
-@testable import QuizEngine
-
-final class Quiz {
-    private let flow: Any
-    
-    private init(flow: Any) {
-        self.flow = flow
-    }
-    
-    static func start<Question, Answer: Equatable, Delegate: QuizDelegate>(questions: [Question], delegate: Delegate, correctAnswers: [Question: Answer]) -> Quiz where Delegate.Question == Question, Delegate.Answer == Answer {
-        let flow = Flow(questions: questions, delegate: delegate, scoring: {
-            scoring($0, correctAnswers: correctAnswers)
-        })
-        flow.start()
-        return Quiz(flow: flow)
-    }
-}
+import QuizEngine
 
 final class QuizTest: XCTestCase {
